Rename package from server to socket
