+++ src/com/thunder_cut/socket/Attachment.java
@@ -3,7 +3,7 @@
  * Author : Arakene
  * Created Date : 2020-01-15
  */
-package com.thunder_cut.server;
+package com.thunder_cut.socket;
 
 import java.nio.ByteBuffer;
 import java.nio.channels.AsynchronousServerSocketChannel;
